ai-engine/requirements.txt
fastapi==0.115.2
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
numpy==1.25.2
python-dotenv==1.0.0
ollama==0.1.0
